+++ blueprints/component-test/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  description: 'Generates a component unit test.'
+};