+++ blueprints/http-mock/index.js
@@ -1,6 +1,12 @@
-var Blueprint = require('../../lib/models/blueprint');
+var EOL = require('os').EOL;
 
-module.exports = Blueprint.extend({
+module.exports = {
+  description: 'Generates a mock api endpoint in /api prefix.',
+
+  anonymousOptions: [
+    'endpoint-path'
+  ],
+
   locals: function(options) {
     return {
       path: '/' + options.entity.name.replace(/^\//, '')
@@ -9,4 +15,4 @@ module.exports = Blueprint.extend({
   afterInstall: function() {
     return this.addPackageToProject('connect-restreamer', '^1.0.0');
   }
-});
+};