+++ blueprints/mixin-test/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  description: 'Generates a mixin unit test.'
+};