+++ blueprints/model-test/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  description: 'Generates a model unit test.'
+};