+++ blueprints/service/index.js
@@ -1,4 +1,3 @@
-var Blueprint = require('../../lib/models/blueprint');
-
-module.exports = Blueprint.extend({
-});
+module.exports = {
+  description: 'Generates a service and initializer for injections.'
+};