+++ tests/fixtures/blueprints/basic/index.js
@@ -1,4 +1,5 @@
 var Blueprint = require('../../../../lib/models/blueprint');
 
 module.exports = Blueprint.extend({
+  description: 'A basic blueprint'
 });