+++ tests/fixtures/blueprints/exporting-object/index.js
@@ -1,3 +1,4 @@
 module.exports = {
+  description: 'A blueprint that exports an object',
   woot: 'someValueHere'
 };